% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{align}
\alias{align}
\title{Sequence Alignment}
\usage{
align(analysis, should_run = FALSE)
}
\arguments{
\item{analysis}{an object of class \code{DifferentialExpressionAnalysis},
containing the choice of workflow and input data}

\item{should_run}{a logical value indicating whether the draft task is to
be executed automatically}
}
\value{
The output is a \code{DifferentialExpressionAnalysis} object,
  holding the same information as the input, but with the task id of the
  alignment task in the \code{alignment_task} slot.
}
\description{
\code{align} creates the alignment task.
}
\examples{
testDEA <- align(testDEA, should_run = TRUE)
}

