% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantificationAndDE.R
\name{analyzeForDE}
\alias{analyzeForDE}
\title{Quantification and Differential Expression Analysis}
\usage{
analyzeForDE(analysis, should_run = FALSE)
}
\arguments{
\item{analysis}{an object of class \code{DifferentialExpressionAnalysis},
containing the choice of workflow and alignment files}

\item{should_run}{a logical value indicating whether the draft task is to
be executed automatically}
}
\value{
The output is a \code{DifferentialExpressionAnalysis} object,
  holding the same information as the input, but with the task id of the
  quantification/DE task in the \code{quantification_DE_task} slot.
}
\description{
\code{analyzeForDE} creates the quantification/DE analysis task.
}
\examples{
testDEA <- analyzeForDE(testDEA, should_run = TRUE)
}

