% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingResults.R
\name{readResults}
\alias{readResults}
\title{Reading Analysis Results}
\usage{
readResults(analysis)
}
\arguments{
\item{analysis}{an object of class \code{DifferentialExpressionAnalysis},
containing the task id of the finished quantification/DE task}
}
\value{
The output is a \code{DifferentialExpressionAnalysis} object,
  holding the same information as the input, but with the analysis results
  contained in the \code{analysis_results} slot.
}
\description{
\code{readResults} is used to read the results once the DE testing is complete.
}
\examples{
testDEA <- readResults(testDEA)
}

