\name{sbde-internal}
\Rdversion{0.1-1}
\alias{lamFn}
\alias{nuFn}
\alias{nuFn.inv}
\alias{sigFn}
\alias{sigFn.inv}
\alias{unitFn}
\alias{sum.sq}
\alias{extract}
\alias{logmean}
\alias{logsum}
\alias{shrinkFn}
\alias{trape}
\alias{klGP}
\alias{proxFn}
\alias{transform.grid}

\title{Internal sbde Functions}
\description{These functions are repeatedly used by the main functions of the package "sbde". Not intended for direct use the by user.}

\keyword{internal}
