% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBDE.R
\docType{package}
\name{SBDE}
\alias{SBDE}
\alias{SBDE-package}
\title{SBDE: Differential Expression Analysis on Seven Bridges Platforms}
\description{
A set of API scripts written with the goal of simplifying comparison of
differentially expressed gene lists produced on Seven Bridges platforms.
}
\section{Vignette}{

LINK!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
}

