% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionAnalysis.R
\docType{class}
\name{DifferentialExpressionAnalysis-class}
\alias{DifferentialExpressionAnalysis-class}
\alias{DifferentialExpressionAnalysis}
\title{S4 Class Representing a DE Analysis}
\description{
Class \code{DifferentialExpressionAnalysis} is a container for holding all the relevant
  info regarding a differential expression analysis on one of Seven Bridges platforms.
}
\section{Slots}{

\describe{
\item{\code{title}}{analysis title}

\item{\code{task}}{id of the chosen DEA task}

\item{\code{results}}{the results of DE analysis}
}}

