% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionAnalysis.R
\name{newDEA}
\alias{newDEA}
\title{New Differential Expression Analysis}
\usage{
newDEA(auth, analysis_title, task_id)
}
\arguments{
\item{auth}{authentication object}

\item{analysis_title}{the title of your analysis}

\item{task_id}{task id}
}
\value{
The output is a new \code{DifferentialExpressionAnalysis} object
}
\description{
\code{newDEA} creates a new object of class \code{DifferentialExpressionAnalysis}.
}
\examples{
library(sevenbridges)
a <- Auth(token = your_token, url = your_platform)
x <- newDEA(a, "Hisat_and_deseq2", "210458f1-0569-4f72-b5d2-0a64ca966c88")
}
